# [Go] Trigger GC Manually